refactor(lang-detect): split franc behind a /franc subpath; upgrade franc-min → franc
Make the main @movar/lang-detect barrel franc-free so importing it never pulls
franc's trigram tables. The franc engine, oracle, and rung-3 resolver move to an
opt-in @movar/lang-detect/franc subpath; classifyBySnippet takes an injected
Rung3Resolver (franc-free by default); the orchestrator dispatcher splits from
the default roster.

Swap franc-min (~82 langs) for full franc (187): it closes the Hebrew gap. Two
short mixed-script English fixtures now abstain (documented in KNOWN_MISSES) —
benign, since English is never blocked; zero Cyrillic regressions, because rung 3
is only:-scoped to the candidates so ru/uk are byte-identical across variants.

The package stays isomorphic — no worker/browser awareness; franc is plain JS.
