+++ packages/lang-detect/src/orchestrator.ts
@@ -1,31 +1,22 @@
 /**
- * Tier-7 orchestrator: walks `ENGINES` in order, returns the first non-null
- * result. No registry API, no confidence floor, no error propagation —
+ * Tier-7 dispatcher: walks an engine roster in order, returns the first
+ * non-null result. No registry API, no confidence floor, no error propagation —
  * engines that throw or return null are equivalent ("fall through").
+ *
+ * Franc-free by construction: this module imports no engines, so importing the
+ * dispatcher never pulls franc. The batteries-included roster + entry point
+ * (`ENGINES`/`detectLanguageFromText`) live in ./default-roster, which imports
+ * the franc engine. A host that supplies its own engines — e.g. the extension's
+ * background-worker franc, reached by message — calls this directly.
  *
  * See docs/on-device-language-detection.md for the design.
  */
 
 import type { DetectContext, DetectedLanguage, LanguageDetectionEngine } from './engine';
-import { chromeAiEngine } from './engines/chrome-ai';
-import { francMinEngine } from './engines/franc-min';
 
-/** Order matters: chrome-ai (Gemini Nano on-device) when available beats
- *  franc-min on accuracy and language coverage. franc-min is the cross-
- *  browser baseline. New engines append. */
-export const ENGINES: readonly LanguageDetectionEngine[] = [chromeAiEngine, francMinEngine];
-
-/** Public surface: detect text language using the built-in engine roster. */
-export async function detectLanguageFromText(
-  text: string,
-  ctx: DetectContext = {},
-): Promise<DetectedLanguage | null> {
-  return detectLanguageFromTextWith(ENGINES, text, ctx);
-}
-
-/** Engine-agnostic dispatcher. Exposed for tests so they can pass synthetic
- *  engine arrays without going through vi.mock. The production
- *  `detectLanguageFromText` is a thin wrapper around this with `ENGINES`. */
+/** Engine-agnostic dispatcher. Tests and alternative hosts pass their own engine
+ *  arrays; the production `detectLanguageFromText` (./default-roster) is a thin
+ *  wrapper around this with the built-in `ENGINES`. */
 export async function detectLanguageFromTextWith(
   engines: readonly LanguageDetectionEngine[],
   text: string,
